Converting data arrays to float32 for efficiency

+++ src/medpipe/pipeline/pipeline.py
@@ -499,6 +499,7 @@ class MedpipePipeline(BaseEstimator, ClassifierMixin):
             X_recal_array = (
                 X_recal.to_numpy() if isinstance(X_recal, pd.DataFrame) else X_recal
             )
+            X_recal_array = X_recal_array.astype(np.float32)
 
         # Convert to numpy if needed
         X_train_array = (
@@ -506,7 +507,11 @@ class MedpipePipeline(BaseEstimator, ClassifierMixin):
         )
         X_test_array = X_test.to_numpy() if isinstance(X_test, pd.DataFrame) else X_test
 
-        return X_train_array, X_test_array, X_recal_array
+        return (
+            X_train_array.astype(np.float32),
+            X_test_array.astype(np.float32),
+            X_recal_array,
+        )
 
     def _get_cv_generator(self) -> StratifiedKFold | GroupKFold:
         """
